pyproject.toml
[project]
name = "har-file-analyzer"
version = "0.1.0"
description = "A tool for parsing HAR files to extract and replay API calls"
authors = [
    {name = "HAR Analyzer Team"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "gunicorn (>=20.1.0,<21.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.1"
pre-commit = "^4.1.0"

